fix: resolve app crashes and improve stability for release v1.0.26
- Fix SQLiteConstraintException by mapping episodeId to chapters during bulk insert.
- Fix NPEs in FeedsViewModel and SettingsViewModel by handling null WorkInfo flows.
- Fix crash in RssParser by adding safe-calls to XML namespace checks.
- Fix SQLiteException by removing PRAGMA busy_timeout from onOpen.
- Remove redundant code conversions and Elvis operators.
- Bump version code to 26.

+++ app/src/main/java/com/yuval/podcasts/utils/LogManager.kt
@@ -34,7 +34,7 @@ class LogManager @Inject constructor(
     private val logDir = File(context.filesDir, "logs").apply { if (!exists()) mkdirs() }
     private val activeLogFile = File(logDir, "app_log_active.jsonl")
     private val previousLogFile = File(logDir, "app_log_previous.jsonl")
-    private val maxFileSize = Constants.MAX_LOG_FILE_SIZE.toLong()
+    private val maxFileSize = Constants.MAX_LOG_FILE_SIZE
 
     private val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS", Locale.US)
     private val logBuffer = mutableListOf<String>()
